created nt_user_check.py
