+++ src/nt_users_tool/nt_user_info.py
@@ -3,7 +3,6 @@ from datetime import date
 from enum import Enum
 
 from nt_users_tool.constants import (
-    NT_USER_MAX_LEN,
     DAYS_EXPIRED_LIMIT,
     DAYS_EXPIRING_15_LIMIT,
     DAYS_EXPIRING_30_LIMIT,
@@ -11,16 +10,6 @@ from nt_users_tool.constants import (
 )
 
 _DATE_NOW = date.today()
-
-
-class InvalidUserNameError(Exception):
-    """Raised for invalid NT user names."""
-
-    def __init__(self, user_name) -> None:
-        self.user_name = user_name
-    
-    def __str__(self) -> str:
-        return f"{self.user_name} is not valid!"
 
 
 class NTUserInfo(NamedTuple):
@@ -60,14 +49,3 @@ def evaluate_user_status(nt_user_info: NTUserInfo) -> NTUserStatus:
         return NTUserStatus.EXPIRING_60_DAYS
     else:
         return NTUserStatus.VALID 
-
-
-def check_nt_user(nt_user:str) -> None:
-    """Check the validity of a user name.
-    
-    :param user_name: user name to check
-    :raises : if the given user name is invalid.
-    """
-    # Check that the length of the user name
-    if len(nt_user) > NT_USER_MAX_LEN:
-        raise IndentationError(nt_user)
